+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### Changes
 * Python: Added OBJECT ENCODING command ([#1471](https://github.com/aws/glide-for-redis/pull/1471))
+* Python: Added OBJECT FREQ command ([#1472](https://github.com/aws/glide-for-redis/pull/1472))
 
 ## 0.4.0 (2024-05-26)